+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 v0.15.10 (??.??.????)
  * new option in form tool for multiple selects
+ * new option to set size for select
 
 v0.15.9 (21.09.2015)
  * new abstract repository class